Move function to its own file
